life-journey-backend/requirements.txt
fastapi>=0.115.4
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.31
pydantic-settings>=2.3.4
pydantic[email]>=2.9.0
python-multipart>=0.0.9
boto3>=1.34.0
httpx>=0.27.0
psycopg[binary]>=3.2.1
alembic>=1.13.2
redis>=5.1.0
celery>=5.4.0
loguru>=0.7.2
mangum>=0.17.0
argon2-cffi>=23.1.0
pyjwt[crypto]>=2.9.0
openai>=1.54.0
slowapi>=0.1.9
